+++ sandbox/consumer-kotlin-multiplatform/linuxMipsel32Main/kotlin/Platform.kt
@@ -1,5 +1,0 @@
-package com.kotlin.multiplatform.sandbox.project
-
-actual object Platform {
-    actual val name: String = "linuxMipsel32"
-}